Spring support

+++ jupiter-common/src/main/java/org/jupiter/common/util/JConstants.java
@@ -61,7 +61,7 @@ public final class JConstants {
     /** 服务默认版本号 */
     public static final String DEFAULT_VERSION = "1.0.0";
     /** 默认的调用超时时间为3秒 **/
-    public static final int DEFAULT_TIMEOUT = SystemPropertyUtil.getInt("jupiter.reader.connection.timeout", 3 * 1000);
+    public static final long DEFAULT_TIMEOUT = SystemPropertyUtil.getInt("jupiter.reader.connection.timeout", 3 * 1000);
     /** 链路read空闲检测 默认60秒, 60秒没读到任何数据会强制关闭连接 */
     public static final int READER_IDLE_TIME_SECONDS = SystemPropertyUtil.getInt("jupiter.reader.idle.time.seconds", 60);
     /** 链路write空闲检测 默认30秒, 30秒没有向链路中写入任何数据时客户端会主动向对端发送心跳 */
